# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Olake is fastest open-source tool for replicating Databases to Apache Iceberg or Data Lakehouse.",
        "title": "Olake UI API",
        "contact": {
            "name": "OLake",
            "url": "https://olake.io/contact/",
            "email": "hello@olake.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/api/v1/project/{projectid}/destinations": {
            "get": {
                "description": "Retrieve a list of all configured destinations within a specific project.",
                "tags": [
                    "Destinations"
                ],
                "summary": "List all destinations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id (default is 123)",
                        "name": "projectid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.DestinationDataItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "failed to validate request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400Response"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401Response"
                        }
                    },
                    "500": {
                        "description": "failed to get destinations",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new destination for a project",
                "tags": [
                    "Destinations"
                ],
                "summary": "Create a new destination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id (default is 123)",
                        "name": "projectid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "destination data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateDestinationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CreateDestinationRequest"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "failed to validate request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400Response"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401Response"
                        }
                    },
                    "413": {
                        "description": "payload too large",
                        "schema": {
                            "$ref": "#/definitions/dto.Error413Response"
                        }
                    },
                    "500": {
                        "description": "failed to create destination",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500Response"
                        }
                    }
                }
            }
        },
        "/api/v1/project/{projectid}/destinations/spec": {
            "post": {
                "description": "Retrieve the UI spec for a specific destination type/version.",
                "tags": [
                    "Destinations"
                ],
                "summary": "Get destination UI spec",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id (default is 123)",
                        "name": "projectid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "spec request data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SpecRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SpecResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "failed to validate request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400Response"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401Response"
                        }
                    },
                    "413": {
                        "description": "payload too large",
                        "schema": {
                            "$ref": "#/definitions/dto.Error413Response"
                        }
                    },
                    "500": {
                        "description": "failed to get spec",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500Response"
                        }
                    }
                }
            }
        },
        "/api/v1/project/{projectid}/destinations/test": {
            "post": {
                "description": "Validate the connection to a destination using the provided configuration details.",
                "tags": [
                    "Destinations"
                ],
                "summary": "Test destination connection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id (default is 123)",
                        "name": "projectid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "test connection data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DestinationTestConnectionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.TestConnectionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "failed to validate request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400Response"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401Response"
                        }
                    },
                    "413": {
                        "description": "payload too large",
                        "schema": {
                            "$ref": "#/definitions/dto.Error413Response"
                        }
                    }
                }
            }
        },
        "/api/v1/project/{projectid}/destinations/versions": {
            "get": {
                "description": "Retrieve the list of available versions for a specific destination connector type.",
                "tags": [
                    "Destinations"
                ],
                "summary": "Get available destination versions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id (default is 123)",
                        "name": "projectid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "destination type",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.VersionsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "failed to validate request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400Response"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401Response"
                        }
                    },
                    "500": {
                        "description": "failed to get versions",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateDestinationRequest": {
            "type": "object",
            "required": [
                "config",
                "name",
                "type",
                "version"
            ],
            "properties": {
                "config": {
                    "type": "string",
                    "example": "{\"catalog_type\":\"glue\",\"warehouse\":\"s3://my-bucket/warehouse\"}"
                },
                "name": {
                    "type": "string",
                    "example": "my-iceberg-destination"
                },
                "type": {
                    "type": "string",
                    "example": "iceberg"
                },
                "version": {
                    "type": "string",
                    "example": "v0.2.7"
                }
            }
        },
...
        "dto.DestinationDataItem": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "string",
                    "example": "{\"catalog_type\":\"glue\",\"warehouse\":\"s3://bucket/warehouse\"}"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "created_by": {
                    "type": "string",
                    "example": "admin"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "jobs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.JobDataItem"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "my-iceberg-destination"
                },
                "type": {
                    "type": "string",
                    "example": "iceberg"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-09T12:00:00Z"
                },
                "updated_by": {
                    "type": "string",
                    "example": "admin"
                },
                "version": {
                    "type": "string",
                    "example": "v0.2.7"
                }
            }
        },
        "dto.DestinationTestConnectionRequest": {
            "type": "object",
            "required": [
                "config",
                "type",
                "version"
            ],
            "properties": {
                "config": {
                    "type": "string",
                    "example": "{\"catalog_type\":\"glue\",\"warehouse\":\"s3://my-bucket/warehouse\"}"
                },
                "source_type": {
                    "type": "string",
                    "example": "postgres"
                },
                "source_version": {
                    "type": "string",
                    "example": "v0.2.7"
                },
                "type": {
                    "type": "string",
                    "example": "iceberg"
                },
                "version": {
                    "type": "string",
                    "example": "v0.2.7"
                }
            }
        },
...
        "dto.Error400Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Bad request"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.Error401Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Authentication required"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "dto.Error413Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Payload too large"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.Error500Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Internal server error"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.JSONResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "operation completed successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.JobDataItem": {
            "type": "object",
            "properties": {
                "activate": {
                    "type": "boolean",
                    "example": true
                },
                "destination_name": {
                    "type": "string",
                    "example": "my-iceberg-destination"
                },
                "destination_type": {
                    "type": "string",
                    "example": "iceberg"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "last_run_state": {
                    "type": "string",
                    "example": "completed"
                },
                "last_run_time": {
                    "type": "string",
                    "example": "2024-01-09T12:00:00Z"
                },
                "name": {
                    "type": "string",
                    "example": "my-sync-job"
                },
                "source_name": {
                    "type": "string",
                    "example": "my-postgres-source"
                },
                "source_type": {
                    "type": "string",
                    "example": "postgres"
                }
            }
        },
...
        "dto.SpecRequest": {
            "type": "object",
            "required": [
                "type",
                "version"
            ],
            "properties": {
                "type": {
                    "description": "enum: postgres,mongodb,mysql,mssql,db2,s3,kafka,iceberg",
                    "type": "string",
                    "example": "postgres"
                },
                "version": {
                    "type": "string",
                    "example": "v0.2.7"
                }
            }
        },
        "dto.SpecResponse": {
            "type": "object",
            "properties": {
                "spec": {
                    "type": "object"
                },
                "type": {
                    "type": "string",
                    "example": "postgres"
                },
                "version": {
                    "type": "string",
                    "example": "v0.2.7"
                }
            }
        },
...
        "dto.TestConnectionResponse": {
            "type": "object",
            "properties": {
                "connection_result": {
                    "type": "object"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                }
            }
        },
...
        "dto.VersionsResponse": {
            "type": "object",
            "properties": {
                "version": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "v0.3.15",
                        "v0.3.14",
                        "v0.3.13",
                        "v0.3.12"
                    ]
                }
            }
        }
    },
...
}